Refactor forms field code
