+++ src/frontend/src/forms/StockForms.tsx
@@ -4,9 +4,9 @@ import {
   ApiFormChangeCallback,
   ApiFormData,
   ApiFormFieldSet
-} from '../../components/forms/fields/ApiFormField';
-import { ApiPaths } from '../../states/ApiState';
-import { openCreateApiForm, openEditApiForm } from '../forms';
+} from '../components/forms/fields/ApiFormField';
+import { openCreateApiForm, openEditApiForm } from '../functions/forms';
+import { ApiPaths } from '../states/ApiState';
 
 /**
  * Construct a set of fields for creating / editing a StockItem instance
